requirements.txt
PyMuPDF==1.23.19
pytesseract==0.3.10
Pillow==10.2.0
python-docx==0.8.11
sentence-transformers==2.2.2
scikit-learn==1.4.1.post1
numpy==1.26.4
huggingface_hub==0.11.1
